requirements.txt
requests
psycopg2
python-dotenv
